requirements.txt
discord.py==1.3.1
PyAutoGUI==0.9.48
python-dotenv==0.11.0
